# Optimize if function when input type is map