--- a/src/Functions/if.cpp
+++ b/src/Functions/if.cpp
@@ -38,7 +38,7 @@ namespace ErrorCodes
     extern const int ILLEGAL_COLUMN;
     extern const int ILLEGAL_TYPE_OF_ARGUMENT;
     extern const int NOT_IMPLEMENTED;
-    extern const int SIZES_OF_ARRAYS_DONT_MATCH;
+    extern const int LOGICAL_ERROR;
 }
 
 namespace