… tests/queries/0_stateless/02974_if_with_map.reference