### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Optimize if function when input type is map, speed up by ~10x

``` sql 
with rand32() % 2 as x select if(x, map(1,2,3,4), map(3,4,5,6)) from numbers(10000000) format Null

Old: 0 rows in set. Elapsed: 4.293 sec. Processed 10.00 million rows, 80.00 MB (2.33 million rows/s., 18.63 MB/s.)
New: 0 rows in set. Elapsed: 0.411 sec. Processed 10.00 million rows, 80.00 MB (24.33 million rows/s., 194.64 MB/s.)
```