@Algunenano 2024-02-01
02906_orc_tuple_field_prune failure is related. Please have a look and add a test once you find a smaller reproducer